Extracted methods for interaction with AddNoteFragment from HomeScreenRobot to HomeAddNoteRobot and introduced addNoteFragment for calling these methods
