+++ app/src/androidTest/java/com/alex/mapnotes/SmokeTests.kt
@@ -73,7 +73,9 @@ class SmokeTests {
         homeScreen {
             isMapDisplayed()
             openAddNote()
-            addNote(noteText)
+            addNoteFragment {
+                addNote(noteText)
+            }
             openSearch()
             searchNoteByText(noteText)
             isNoteInSearchHasResults(noteText)
