+++ app/src/androidTest/java/com/alex/mapnotes/home/HomeActivityTest.kt
@@ -46,7 +46,9 @@ class HomeActivityTest : MockTest() {
         homeScreen {
             displayAsEntryPoint()
             openAddNote()
-            isSuccessfullyDisplayedAddNote()
+            addNoteFragment {
+                isSuccessfullyDisplayedAddNote()
+            }
         }
     }
